Step 12
The method attribute specifies how to send form-data to the URL specified in the action attribute. The form-data can be sent via a GET request as URL parameters (with method=get) or via a POST request as data in the request body (with method=post).

Set the method attribute to send your form data via a POST request.