Step 36
Submitting the form with an option selected would not send a useful value to the server. As such, each option needs to be given a value attribute. Without which, the text content of the option will be submitted to the server.

Give the first option a value of 'double quote', and the subsequent option elements value attributes from 1 to 4.